+++ packages/core/test/mocks/backend.ts
@@ -1,5 +1,5 @@
 import { Session } from '@sentry/hub';
-import { Event, Options, SeverityLevel, Transport } from '@sentry/types';
+import { Event, Options, Severity, Transport } from '@sentry/types';
 import { resolvedSyncPromise } from '@sentry/utils';
 
 import { BaseBackend } from '../../src/basebackend';
@@ -38,7 +38,7 @@ export class TestBackend extends BaseBackend<TestOptions> {
     });
   }
 
-  public eventFromMessage(message: string, level: SeverityLevel = 'info'): PromiseLike<Event> {
+  public eventFromMessage(message: string, level: Severity = Severity.Info): PromiseLike<Event> {
     return resolvedSyncPromise({ message, level });
   }