+++ packages/nextjs/src/config/webpack.ts
@@ -77,12 +77,17 @@ export function constructWebpackConfigFunction(
     if (enableWebpackPlugin) {
       // TODO Handle possibility that user is using `SourceMapDevToolPlugin` (see
       // https://webpack.js.org/plugins/source-map-dev-tool-plugin/)
-      // TODO Give user option to use `hidden-source-map` ?
 
-      // Next doesn't let you change this is dev even if you want to - see
+      // Next doesn't let you change `devtool` in dev even if you want to, so don't bother trying - see
       // https://github.com/vercel/next.js/blob/master/errors/improper-devtool.md
       if (!buildContext.dev) {
-        newConfig.devtool = 'source-map';
+        // `hidden-source-map` produces the same sourcemaps as `source-map`, but doesn't include the `sourceMappingURL`
+        // comment at the bottom. For folks who aren't publicly hosting their sourcemaps, this is helpful because then
+        // the browser won't look for them and throw errors into the console when it can't find them. Because this is a
+        // front-end-only problem, and because `sentry-cli` handles sourcemaps more reliably with the comment than
+        // without, the option to use `hidden-source-map` only applies to the client-side build.
+        newConfig.devtool =
+          userNextConfig.sentry?.hideSourceMaps && !buildContext.isServer ? 'hidden-source-map' : 'source-map';
       }
 
       newConfig.plugins = newConfig.plugins || [];
@@ -116,13 +121,13 @@ async function addSentryToEntryProperty(
   // we know is that it won't have gotten *simpler* in form, so we only need to worry about the object and function
   // options. See https://webpack.js.org/configuration/entry-context/#entry.
 
+  const { isServer, dir: projectDir, dev: isDev, config: userNextConfig } = buildContext;
+
   const newEntryProperty =
     typeof currentEntryProperty === 'function' ? await currentEntryProperty() : { ...currentEntryProperty };
 
   // `sentry.server.config.js` or `sentry.client.config.js` (or their TS equivalents)
-  const userConfigFile = buildContext.isServer
-    ? getUserConfigFile(buildContext.dir, 'server')
-    : getUserConfigFile(buildContext.dir, 'client');
+  const userConfigFile = isServer ? getUserConfigFile(projectDir, 'server') : getUserConfigFile(projectDir, 'client');
 
   // we need to turn the filename into a path so webpack can find it
   const filesToInject = [`./${userConfigFile}`];
@@ -131,12 +136,12 @@ async function addSentryToEntryProperty(
   // server SDK's default `RewriteFrames` instance (which needs it at runtime). Doesn't work when using the dev server
   // because it somehow tricks the file watcher into thinking that compilation itself is a file change, triggering an
   // infinite recompiling loop. (This should be fine because we don't upload sourcemaps in dev in any case.)
-  if (buildContext.isServer && !buildContext.dev) {
+  if (isServer && !isDev) {
     const rewriteFramesHelper = path.resolve(
       fs.mkdtempSync(path.resolve(os.tmpdir(), 'sentry-')),
       'rewriteFramesHelper.js',
     );
-    fs.writeFileSync(rewriteFramesHelper, `global.__rewriteFramesDistDir__ = '${buildContext.config.distDir}';\n`);
+    fs.writeFileSync(rewriteFramesHelper, `global.__rewriteFramesDistDir__ = '${userNextConfig.distDir}';\n`);
     // stick our helper file ahead of the user's config file so the value is in the global namespace *before*
     // `Sentry.init()` is called
     filesToInject.unshift(rewriteFramesHelper);
@@ -144,7 +149,7 @@ async function addSentryToEntryProperty(
 
   // inject into all entry points which might contain user's code
   for (const entryPointName in newEntryProperty) {
-    if (shouldAddSentryToEntryPoint(entryPointName)) {
+    if (shouldAddSentryToEntryPoint(entryPointName, isServer)) {
       addFilesToExistingEntryPoint(newEntryProperty, entryPointName, filesToInject);
     }
   }
@@ -251,10 +256,15 @@ function checkWebpackPluginOverrides(
  * Determine if this is an entry point into which both `Sentry.init()` code and the release value should be injected
  *
  * @param entryPointName The name of the entry point in question
+ * @param isServer Whether or not this function is being called in the context of a server build
  * @returns `true` if sentry code should be injected, and `false` otherwise
  */
-function shouldAddSentryToEntryPoint(entryPointName: string): boolean {
-  return entryPointName === 'pages/_app' || entryPointName.includes('pages/api');
+function shouldAddSentryToEntryPoint(entryPointName: string, isServer: boolean): boolean {
+  return (
+    entryPointName === 'pages/_app' ||
+    entryPointName.includes('pages/api') ||
+    (isServer && entryPointName === 'pages/_error')
+  );
 }
 
 /**
@@ -269,13 +279,13 @@ export function getWebpackPluginOptions(
   buildContext: BuildContext,
   userPluginOptions: Partial<SentryWebpackPluginOptions>,
 ): SentryWebpackPluginOptions {
-  const { isServer, dir: projectDir, buildId, dev: isDev, config: nextConfig, webpack } = buildContext;
-  const distDir = nextConfig.distDir ?? '.next'; // `.next` is the default directory
+  const { buildId, isServer, webpack, config: userNextConfig, dev: isDev, dir: projectDir } = buildContext;
+  const distDir = userNextConfig.distDir ?? '.next'; // `.next` is the default directory
 
   const isWebpack5 = webpack.version.startsWith('5');
-  const isServerless = nextConfig.target === 'experimental-serverless-trace';
+  const isServerless = userNextConfig.target === 'experimental-serverless-trace';
   const hasSentryProperties = fs.existsSync(path.resolve(projectDir, 'sentry.properties'));
-  const urlPrefix = nextConfig.basePath ? `~${nextConfig.basePath}/_next` : '~/_next';
+  const urlPrefix = userNextConfig.basePath ? `~${userNextConfig.basePath}/_next` : '~/_next';
 
   const serverInclude = isServerless
     ? [{ paths: [`${distDir}/serverless/`], urlPrefix: `${urlPrefix}/serverless` }]
@@ -295,7 +305,7 @@ export function getWebpackPluginOptions(
     configFile: hasSentryProperties ? 'sentry.properties' : undefined,
     stripPrefix: ['webpack://_N_E/'],
     urlPrefix,
-    entries: shouldAddSentryToEntryPoint,
+    entries: (entryPointName: string) => shouldAddSentryToEntryPoint(entryPointName, isServer),
     release: getSentryRelease(buildId),
     dryRun: isDev,
   });