+++ packages/vue/src/components.ts
@@ -71,10 +71,12 @@ export const generateComponentTrace = (vm?: ViewModel): string => {
     const formattedTree = tree
       .map(
         (vm, i) =>
-          `${(i === 0 ? '---> ' : repeat(' ', 5 + i * 2)) +
+          `${
+            (i === 0 ? '---> ' : repeat(' ', 5 + i * 2)) +
             (Array.isArray(vm)
               ? `${formatComponentName(vm[0])}... (${vm[1]} recursive calls)`
-              : formatComponentName(vm))}`,
+              : formatComponentName(vm))
+          }`,
       )
       .join('\n');